Added an ambulance
Added an ambulance that is rushing from behind. I dont know if this is just me being unlucky but it ive tested a few times and the ambulance spawns really close to the start of playing the game and then eventually it like doesnt spawn for a while but idk if its me being unlucky. May need to have a system that makes the chances fo something spawning fairly equal to everything else